# optimize `initcap` function by avoiding memory allocation