@waynexia 2025-07-28
Thank you @alamb.

>[docs.rs/arrow/latest/arrow/array/type.StringViewBuilder.html#method.append_block](https://docs.rs/arrow/latest/arrow/array/type.StringViewBuilder.html#method.append_block)

looks like this still requires some intermediate container 🤔 From current APIs this seems unavoidable. Unless we maintain the logic of creating view element in this function, which doesn't seem to be a good idea. But I'm wondering if we can wrap something like `ViewItemBuilder` that holds the similar logic as [here](https://docs.rs/arrow-array/55.2.0/src/arrow_array/builder/generic_bytes_view_builder.rs.html#315) and implement `std::io::Write` (or can be pushed/extended byte-by-byte) like `StringArrayBuilder` so that we don't need an intermediate container anymore?